@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { Theme } from '@twilio-paste/core/theme';
-import { Routes } from './routes';
+import { AppRoutes } from './routes';
 import { GlobalContext } from './state/context';
 import { ThemeMode } from './types';
 
@@ -11,7 +11,7 @@ export const App = () => {
   return (
     <GlobalContext.Provider value={globalContextValue}>
       <Theme.Provider theme={isDarkMode ? ThemeMode.Dark : ThemeMode.Default}>
-        <Routes />
+        <AppRoutes />
       </Theme.Provider>
     </GlobalContext.Provider>
   );
