@@ -1,10 +1,7 @@
 import { useParams } from 'react-router-dom';
 
 export const Home = () => {
-  interface Params {
-    account: string;
-  }
-  const { account } = useParams<Params>();
+  const { account } = useParams();
 
   return <>{account}</>;
 };
